requirements.txt
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.0
torch==2.4.1
torchvision==0.19.1
einops==0.8.0
resize-right==0.0.2
tqdm
numpy
pyav==13.0.0
pandas==2.2.3
pyarrow==17.0.0
plotly
